Extract PageButton
